old-tests/submission/Opera/preload/metadata/preload-metadata-to-none-event-order-ogg.html

<!doctype html>
<!-- This file was generated from 'preload-metadata-to-none-event-order.tpl' using 'tpl2tc.py'. Please do not edit manually! -->
<title>HTMLMediaElement.preload:metadata event order when preload is set to none after loadstart is fired - ogg</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../resources/common.js"></script>
<p id="log">FAIL (script didn't run)</p>
<audio preload="metadata" controls></audio>
<video preload="metadata" controls></video>
<script>
var tests = init_tests("HTMLAudioElement.preload:metadata event order when preload is set to none after loadstart is fired - ogg", "HTMLVideoElement.preload:metadata event order when preload is set to none after loadstart is fired - ogg", {timeout:10000});
tests.forEach(function(vars) {
    var t = vars[0];
    t.step(function() {
	var tag_name = vars[1];
	var events_actual = '';
	var node = document.getElementsByTagName(tag_name)[0];
	var events  = ['loadstart', 'progress', 'suspend', 'abort', 'error', 'loadedmetadata', 'loadeddata', 'canplay', 'canplaythrough'];
	for (var e in events) {
	    node.addEventListener(events[e], handleEvent, false);
	}
	node.addEventListener('loadstart', startTest, false);
	node.addEventListener('error', function() {
	    if (!node.canPlayType(tag_name + '/ogg')) {
		t.step(function() {
		    assert_unreached("Cannot play '" + tag_name + "/ogg'");
		    t.done();
		});
	    }
	}, false);
	node.src = '../resources/preload.ogv?pipe=trickle(100000:d1:r2)&nocache=' + Math.random();

	function startTest(e) {
	    node.removeEventListener('loadstart', startTest, false);
	    node.preload = 'none';
	}

	function handleEvent(e) {
	    events_actual += e.type + ' ';
	    if (e.type == 'suspend') endTest();
	}

	function endTest() {
	    t.step(function() {
		assert_regexp_match(events_actual, /^loadstart (progress )+loadedmetadata (progress )*loadeddata (progress )*suspend $/g, 'Event order check');
		assert_equals(node.preload, 'none', "'preload' value should be 'none'");
		t.done();
	    });
	}
    });
});
</script>
